ROC-1183 - Separated dialectRT_test.go into two files so that tests can be exported and called in rocos-agent.
